Shifted our sponsors into a seperate component
