+++ src/app/(pages)/sponsors/page.tsx
@@ -1,14 +1,7 @@
 import React from "react";
-import SilverSponsorCard from "@/components/SilverSponsorCard";
-import GoldSponsorCard from "@/components/GoldSponsorCard";
+import OurSponsors from "@/components/OurSponsors";
 
-interface Sponsor {
-  name: string;
-  logo: string;
-  link: string;
-}
-
-const goldSponsors: Sponsor[] = [
+const goldSponsors = [
   {
     name: "Jane Street",
     logo: "sponsors-images/jane-st.png",
@@ -16,7 +9,7 @@ const goldSponsors: Sponsor[] = [
   },
 ];
 
-const silverSponsors: Sponsor[] = [
+const silverSponsors = [
   {
     name: "IMC Trading",
     logo: "sponsors-images/IMC.png",
@@ -58,38 +51,12 @@ const SponsorsPage: React.FC = () => {
   return (
     <div className="page-padding bg-gray-900">
       <h1 className="page-title text-center">Our 2024 Sponsors</h1>
-      <div className="mt-8">
-        <h2 className="text-2xl font-bold text-center mb-4">Gold Sponsors</h2>
-        <div className="mx-auto w-20 my-2 bg-accent h-[2px]"></div>
-        <div className="flex flex-wrap justify-center gap-4 max-w-6xl mx-auto">
-          {goldSponsors.map((sponsor, index) => (
-            <div key={index} className="px-12 py-2">
-              <GoldSponsorCard
-                key={index}
-                name={sponsor.name}
-                logo={sponsor.logo}
-                link={sponsor.link}
-              />
-            </div>
-          ))}
-        </div>
-        <h2 className="text-2xl font-bold text-center mt-16 mb-4">
-          Silver Sponsors
-        </h2>
-        <div className="mx-auto w-20 my-2 bg-accent h-[2px]"></div>
-        <div className="flex flex-wrap justify-center gap-4 max-w-6xl mx-auto">
-          {silverSponsors.map((sponsor, index) => (
-            <div key={index} className="px-12 py-2">
-              <SilverSponsorCard
-                key={index}
-                name={sponsor.name}
-                logo={sponsor.logo}
-                link={sponsor.link}
-              />
-            </div>
-          ))}
-        </div>
-      </div>
+      <OurSponsors title="Gold Sponsors" sponsors={goldSponsors} type="gold" />
+      <OurSponsors
+        title="Silver Sponsors"
+        sponsors={silverSponsors}
+        type="silver"
+      />
     </div>
   );
 };
